+++ src/index.test.ts
@@ -1,17 +1,22 @@
 import type { UnstableDevWorker } from "wrangler";
 import "jest-extended";
-import { unstable_dev } from "wrangler";
 import { repo, title, version } from "./meta";
+import { unstable_dev } from "wrangler";
+import fetchMock from "jest-fetch-mock";
+fetchMock.enableMocks(); // Enables use of `Request` and `Response` objects
 
 type Response = Awaited<ReturnType<UnstableDevWorker["fetch"]>>;
 
-function expectHeaders(response: Response): void {
+function expectHeaders(response: Response | globalThis.Response): void {
 	expect(response).toHaveProperty("headers");
 	expect(response.headers.get("Vary")).toBe("*");
 	expect(response.headers.get("Cache-Control")).toBe("no-store");
 	expect(response.headers.get("Access-Control-Allow-Origin")).toBe("*");
 	expect(response.headers.get("Access-Control-Allow-Methods")).toInclude("GET");
+	expect(response.headers.get("Access-Control-Allow-Methods")).toInclude("HEAD");
+	expect(response.headers.get("Access-Control-Allow-Methods")).toInclude("OPTIONS");
 	expect(response.headers.get("Access-Control-Allow-Headers")).toInclude("Accept");
+	expect(response.headers.get("X-Clacks-Overhead")).toBeString(); // don't care what we put here
 }
 
 describe("IP Echo", () => {
@@ -25,6 +30,7 @@ describe("IP Echo", () => {
 		// `beforeEach` would spam macOS testers with "Can Node use the network?" prompts
 		worker = await unstable_dev("src/index.ts", {
 			experimental: { disableExperimentalWarning: true },
+			vars: { NODE_ENV: "test" }, // match Jest's Node behavior
 		});
 	});
 
@@ -32,7 +38,50 @@ describe("IP Echo", () => {
 		await worker.stop();
 	});
 
-	// TODO: Mock `import { fetchHandler as fetch } from "fetchHandler";` so we don't duplicate tests here
+	describe("routing", () => {
+		test("returns 404 for unknown route", async () => {
+			const res = await worker.fetch(new URL("lolz", url));
+			expect(res.headers.get("Content-Type")).toBe("text/plain;charset=UTF-8");
+			expect(res.status).toBe(404);
+			expectHeaders(res);
+		});
+
+		test("sanity test is invisible in prod", async () => {
+			const worker = await unstable_dev("src/index.ts", {
+				experimental: { disableExperimentalWarning: true },
+				port: 60546, // eslint-disable-line unicorn/numeric-separators-style
+				// omit NODE_ENV
+			});
+			try {
+				const res = await worker.fetch(new URL("test-error", url));
+				expect(res.headers.get("Content-Type")).toBe("text/plain;charset=UTF-8");
+				expect(res.status).toBe(404);
+				expect(await res.text()).toBe("Not found\n");
+				expectHeaders(res);
+			} finally {
+				await worker.stop();
+			}
+		});
+
+		test("returns 500 and text if handler throws unknown error", async () => {
+			const res = await worker.fetch(new URL("test-error", url));
+			expect(res.headers.get("Content-Type")).toBe("text/plain;charset=UTF-8");
+			expect(res.status).toBe(500);
+			expect(await res.text()).toBe("Internal error\n");
+			expectHeaders(res);
+		});
+
+		test("returns 500 and JSON if handler throws unknown error", async () => {
+			const res = await worker.fetch(new URL("test-error", url), {
+				headers: { Accept: "application/json" },
+			});
+			expect(res.headers.get("Content-Type")).toBe("application/json;charset=UTF-8");
+			expect(res.status).toBe(500);
+			expect(await res.json()).toStrictEqual({ status: 500, message: "Internal error" });
+			expectHeaders(res);
+		});
+	});
+
 	describe("echo", () => {
 		test("responds correctly to CORS preflight", async () => {
 			const response = await worker.fetch(url, { method: "OPTIONS" });
@@ -42,20 +91,31 @@ describe("IP Echo", () => {
 		});
 
 		// See https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods; note that TRACE is not supported
-		const BadMethods = ["HEAD", "POST", "PUT", "DELETE", "PATCH"] as const;
+		const BadMethods = ["POST", "PUT", "DELETE", "PATCH"] as const;
 		test.each(BadMethods)("responds 405 to %s requests", async method => {
 			const response = await worker.fetch(url, { method });
 			expectHeaders(response);
 			expect(response.status).toBe(405);
-			expect(await response.text()).toBe("");
+			expect(await response.text()).toBe("Not Allowed\n");
 		});
 
-		test("responds with the caller's IP address", async () => {
+		test("GET responds with the caller's IP address", async () => {
 			const response = await worker.fetch(url, { headers: { [IP_HEADER_NAME]: TEST_IP } });
-			expect(response.headers.get("Content-Type")).toBe("text/plain");
-
+			expectHeaders(response);
+			expect(response.headers.get("Content-Type")).toBe("text/plain;charset=UTF-8");
 			expect(response.status).toBe(200);
 			expect(await response.text()).toBe(TEST_IP.concat("\n"));
+		});
+
+		test("HEAD responds with appropriate default headers", async () => {
+			const response = await worker.fetch(url, {
+				method: "HEAD",
+				headers: { [IP_HEADER_NAME]: TEST_IP },
+			});
+			expectHeaders(response);
+			expect(response.headers.get("Content-Type")).toBe("text/plain;charset=UTF-8");
+			expect(response.status).toBe(200);
+			expect(await response.text()).toBe("");
 		});
 
 		test("responds with the caller's IP address in JSON", async () => {
@@ -65,10 +125,24 @@ describe("IP Echo", () => {
 					Accept: "application/json",
 				},
 			});
-			expect(response.headers.get("Content-Type")).toBe("application/json");
+			expect(response.headers.get("Content-Type")).toBe("application/json;charset=UTF-8");
 			expectHeaders(response);
 			expect(response.status).toBe(200);
 			expect(await response.json()).toBe(TEST_IP);
+		});
+
+		test("HEAD responds with appropriate headers", async () => {
+			const response = await worker.fetch(url, {
+				method: "HEAD",
+				headers: {
+					[IP_HEADER_NAME]: TEST_IP,
+					Accept: "application/json",
+				},
+			});
+			expect(response.headers.get("Content-Type")).toBe("application/json;charset=UTF-8");
+			expectHeaders(response);
+			expect(response.status).toBe(200);
+			expect(await response.text()).toBe("");
 		});
 
 		test("responds with 500 if there is no 'CF-Connecting-IP' header", async () => {
@@ -80,13 +154,14 @@ describe("IP Echo", () => {
 					Accept: "application/json",
 				},
 			});
+			expectHeaders(response);
 			expect(await response.json()).toStrictEqual({ status: 500, message: "Internal error" });
 			expect(response.status).toBe(500);
 		});
 
 		test("responds with 404 if the request path is not root", async () => {
 			const response = await worker.fetch(new URL("test", url));
-			expect(response.headers.get("Content-Type")).toBe("text/plain");
+			expect(response.headers.get("Content-Type")).toBe("text/plain;charset=UTF-8");
 			expectHeaders(response);
 			expect(response.status).toBe(404);
 			expect(await response.text()).toBe("Not found\n");
@@ -96,22 +171,44 @@ describe("IP Echo", () => {
 			const response = await worker.fetch(new URL("test", url), {
 				headers: { Accept: "application/json" },
 			});
-			expect(response.headers.get("Content-Type")).toBe("application/json");
+			expect(response.headers.get("Content-Type")).toBe("application/json;charset=UTF-8");
 			expectHeaders(response);
 			expect(response.status).toBe(404);
 			expect(await response.json()).toStrictEqual({ status: 404, message: "Not found" });
 		});
-
-		// TODO: Test that it responds with 500 if the request is to an invalid URL (should never happen, but we catch it anyway)
 	});
 
 	describe("about", () => {
+		test("responds correctly to CORS preflight", async () => {
+			const response = await worker.fetch(new URL("about", url), { method: "OPTIONS" });
+			expectHeaders(response);
+			expect(response.status).toBe(204);
+			expect(response.body).toBe(null);
+		});
+
+		// See https://developer.mozilla.org/en-US/docs/Web/HTTP/Methods; note that TRACE is not supported
+		const BadMethods = ["POST", "PUT", "DELETE", "PATCH"] as const;
+		test.each(BadMethods)("responds 405 to %s requests", async method => {
+			const response = await worker.fetch(new URL("about", url), { method });
+			expectHeaders(response);
+			expect(response.status).toBe(405);
+			expect(await response.text()).toBe("Not Allowed\n");
+		});
+
 		test("responds with appropriate metadata", async () => {
 			const response = await worker.fetch(new URL("about", url));
-			expect(response.headers.get("Content-Type")).toBe("application/json");
+			expect(response.headers.get("Content-Type")).toBe("application/json;charset=UTF-8");
 			expectHeaders(response);
 			expect(response.status).toBe(200);
 			expect(await response.json()).toStrictEqual({ repo, title, version });
+		});
+
+		test("HEAD responds with appropriate headers", async () => {
+			const response = await worker.fetch(new URL("about", url), { method: "HEAD" });
+			expect(response.headers.get("Content-Type")).toBe("application/json;charset=UTF-8");
+			expectHeaders(response);
+			expect(response.status).toBe(200);
+			expect(await response.text()).toStrictEqual("");
 		});
 	});
 });