+++ src/routes/echo.test.ts
@@ -1,3 +1,4 @@
+import type { Context } from "hono";
 import { echo } from "./echo";
 import { InternalError } from "../errors/InternalError";
 import fetchMock from "jest-fetch-mock";
@@ -12,11 +13,13 @@ describe("echo", () => {
 		const req = new Request(url, {
 			headers: { [IP_HEADER_NAME]: TEST_IP },
 		});
-		expect(echo(req)).toBe(TEST_IP);
+		const c = { req } as unknown as Context;
+		expect(echo(c)).toBe(TEST_IP);
 	});
 
 	test("throws if IP is not found", () => {
 		const req = new Request(url);
-		expect(() => echo(req)).toThrow(InternalError);
+		const c = { req } as unknown as Context;
+		expect(() => echo(c)).toThrow(InternalError);
 	});
 });