+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- `Content-Type` response headers now include a `charset` declaration.
+
+### Fixed
+
+- `HEAD` requests now [correctly](https://developer.mozilla.org/en-US/docs/web/http/methods/head) respond with the headers that a `GET` request might have sent.
+
 ## [1.1.0] - 2023-03-12
 
 ### Added
@@ -35,6 +45,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Cloudflare Worker to echo the caller's IP address.
 
+[Unreleased]: https://github.com/AverageHelper/ip-echo-cloudflare/compare/v1.1.0...HEAD
 [1.1.0]: https://github.com/AverageHelper/ip-echo-cloudflare/compare/v1.0.3...v1.1.0
 [1.0.3]: https://github.com/AverageHelper/ip-echo-cloudflare/compare/v1.0.2...v1.0.3
 [1.0.2]: https://github.com/AverageHelper/ip-echo-cloudflare/compare/v1.0.1...v1.0.2